+++ README.md
@@ -3,10 +3,10 @@ _Le bot Telegram de la [CLIC](https://clic.epfl.ch)_
 
 ### Requirements
 
-Install the right [Telegram API wrapper](https://github.com/python-telegram-bot/python-telegram-bot) for Python 3.8. Namely
+Install dependencies [Telegram API wrapper](https://github.com/python-telegram-bot/python-telegram-bot) and BeautifulSoup 4
 
 ```
-pip install python-telegram-bot==3.14
+pip install -r reqs.txt
 ```
 
 See the [Telegram Bot API documentation](https://core.telegram.org/bots) to obtain your bot token, to be stored in a ```.keys``` JSON file under ```{ "token": "YOUR_TOKEN" }```. Then simply launch the bot
