+++ reqs.txt
@@ -1,0 +1,2 @@
+python-telegram-bot==13.4
+bs4
