+++ robot/config/soup.sh
