+++ robot/handlers/__init__.py
@@ -1,0 +1,9 @@
+from .birthday import birthday
+from .coffee import coffee
+from .desk import desk
+from .help import help
+from .jul import jul
+from .poll import poll_conv_handler, poll_keyboard_handler, stats
+from .soup import soup
+from .text import hugo, arthur, rayan, reuf, noel
+from .year import year
